daos: remove prefix from docs and update call to duns_resolve_path
This just removes the daos-prefix option from the documentation
because it is no longer necessary after a recent DAOS PR
that allows passing a subset of a UNS path to duns_resolve_path.
The option is removed from the documentation but not removed
from the code in order to preserve backwards compatibility with
the DAOS tests for the datamover.

The call to duns_resolve_path was updated in the case that the
last directory or file in the UNS path does not yet exist, since
it doesn't recurse the rest of the paths in the directory tree
in that case.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>